+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - User profile page to create API keys
 - Show message when reaching inbox zero
 - Add notification kind filtering
+- Display Linear notification reason
 
 ### Changed